Added WriteGoModFiles and Mock* funcs to testhelper
